backend/requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
mysql-connector-python>=8.1.0
numpy<2
pandas==2.0.3
openpyxl>=3.1.2
python-dotenv>=1.0.0
requests>=2.31.0
SpeechRecognition>=3.10.0
pydub>=0.25.1
matplotlib==3.7.5
seaborn==0.12.2